edited

+++ Example_5_Task5/Program.cs
@@ -11,5 +11,5 @@ using System.Text;
 
 Console.OutputEncoding = Encoding.UTF8;
 Console.Write("Введите трёхзначное число: ");
-float n = Convert.ToInt32(Console.ReadLine());
+double n = Convert.ToDouble(Console.ReadLine());
 Console.WriteLine(n % 10);
